requirements.txt
streamlit==1.29.0
aiohttp==3.9.1
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
openai==1.3.7
anthropic==0.8.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
stripe==7.8.0
pandas==2.1.3
plotly==5.17.0
Pillow==10.1.0
requests==2.31.0
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
